noimplicityany changes for files operation,service,service-object from src folder and operation,service,service-object from test and tsconfig file

+++ test/service-object.ts
@@ -18,10 +18,11 @@ import * as pfy from '@google-cloud/promisify';
 import * as assert from 'assert';
 import * as extend from 'extend';
 import * as proxyquire from 'proxyquire';
+import {GrpcServiceObject} from '../src/service-object';
 
 let promisified = false;
 const fakePfy = Object.assign({}, pfy, {
-  promisifyAll(klass) {
+  promisifyAll(klass: {name: string}) {
     if (klass.name === 'GrpcServiceObject') {
       promisified = true;
     }
@@ -35,26 +36,42 @@ class FakeServiceObject {
   }
 }
 
+interface GrpcServiceObjectInterface {
+  methods?: {
+    delete: {protoOpts: {}, reqOpts: {}};
+    getMetadata: {protoOpts: {}, reqOpts: {}};
+    setMetadata: {reqOpts: {}, protoOpts: {}}
+  };
+  request?: Function;
+  delete?: Function;
+  getMetadata?: Function;
+  metadata?: {};
+  setMetadata?: Function;
+  parent?: {};
+  requestStream?: Function;
+  requestWritableStream?: Function;
+}
+
 describe('GrpcServiceObject', () => {
   // tslint:disable-next-line:variable-name
-  let GrpcServiceObject;
-  let grpcServiceObject;
+  let GrpcServiceObjectType: typeof GrpcServiceObject;
+  let grpcServiceObject: GrpcServiceObjectInterface;
 
   const CONFIG = {};
   const PROTO_OPTS = {};
   const REQ_OPTS = {};
 
   before(() => {
-    GrpcServiceObject = proxyquire('../src/service-object', {
-                          '@google-cloud/common': {
-                            ServiceObject: FakeServiceObject,
-                          },
-                          '@google-cloud/promisify': fakePfy
-                        }).GrpcServiceObject;
+    GrpcServiceObjectType = proxyquire('../src/service-object', {
+                              '@google-cloud/common': {
+                                ServiceObject: FakeServiceObject,
+                              },
+                              '@google-cloud/promisify': fakePfy
+                            }).GrpcServiceObject;
   });
 
   beforeEach(() => {
-    grpcServiceObject = new GrpcServiceObject(CONFIG);
+    grpcServiceObject = new GrpcServiceObjectType(CONFIG) as {};
 
     grpcServiceObject.methods = {
       delete: {
@@ -76,7 +93,8 @@ describe('GrpcServiceObject', () => {
     it('should inherit from ServiceObject', () => {
       assert(grpcServiceObject instanceof FakeServiceObject);
 
-      const calledWith = grpcServiceObject.calledWith_;
+      const calledWith =
+          (grpcServiceObject as {calledWith_: Array<{}>}).calledWith_;
       assert.strictEqual(calledWith[0], CONFIG);
     });
 
@@ -87,36 +105,54 @@ describe('GrpcServiceObject', () => {
 
   describe('delete', () => {
     it('should make the correct request', done => {
-      grpcServiceObject.request = (protoOpts, reqOpts, callback) => {
-        const deleteMethod = grpcServiceObject.methods.delete;
-        assert.strictEqual(protoOpts, deleteMethod.protoOpts);
-        assert.strictEqual(reqOpts, deleteMethod.reqOpts);
-        callback();  // done()
-      };
-
-      grpcServiceObject.delete(done);
+      grpcServiceObject.request =
+          (protoOpts: {}, reqOpts: {}, callback: Function) => {
+            const deleteMethod = grpcServiceObject.methods ?
+                grpcServiceObject.methods.delete :
+                undefined;
+            assert.strictEqual(deleteMethod !== undefined, true);
+            if (deleteMethod) {
+              assert.strictEqual(protoOpts, deleteMethod.protoOpts);
+              assert.strictEqual(reqOpts, deleteMethod.reqOpts);
+              callback();  // done()
+            }
+          };
+      if (grpcServiceObject.delete) {
+        grpcServiceObject.delete(done);
+      }
     });
 
     it('should not require a callback', done => {
-      grpcServiceObject.request = (protoOpts, reqOpts, callback) => {
-        assert.doesNotThrow(callback);
-        done();
-      };
+      grpcServiceObject.request =
+          (protoOpts: {}, reqOpts: {}, callback: Function) => {
+            assert.doesNotThrow(callback);
+            done();
+          };
 
-      grpcServiceObject.delete();
+      if (grpcServiceObject.delete) {
+        grpcServiceObject.delete();
+      }
     });
   });
 
   describe('getMetadata', () => {
     it('should make the correct request', done => {
-      grpcServiceObject.request = (protoOpts, reqOpts, callback) => {
-        const getMetadataMethod = grpcServiceObject.methods.getMetadata;
-        assert.strictEqual(protoOpts, getMetadataMethod.protoOpts);
-        assert.strictEqual(reqOpts, getMetadataMethod.reqOpts);
-        callback();  // done()
-      };
+      grpcServiceObject.request =
+          (protoOpts: {}, reqOpts: {}, callback: Function) => {
+            const getMetadataMethod = grpcServiceObject.methods ?
+                grpcServiceObject.methods.getMetadata :
+                undefined;
+            assert.strictEqual(getMetadataMethod !== undefined, true);
+            if (getMetadataMethod) {
+              assert.strictEqual(protoOpts, getMetadataMethod.protoOpts);
+              assert.strictEqual(reqOpts, getMetadataMethod.reqOpts);
+              callback();  // done()
+            }
+          };
 
-      grpcServiceObject.getMetadata(done);
+      if (grpcServiceObject.getMetadata) {
+        grpcServiceObject.getMetadata(done);
+      }
     });
 
     describe('error', () => {
@@ -124,18 +160,23 @@ describe('GrpcServiceObject', () => {
       const apiResponse = {};
 
       beforeEach(() => {
-        grpcServiceObject.request = (protoOpts, reqOpts, callback) => {
-          callback(error, apiResponse);
-        };
+        grpcServiceObject.request =
+            (protoOpts: {}, reqOpts: {}, callback: Function) => {
+              callback(error, apiResponse);
+            };
       });
 
       it('should execute callback with error & API response', done => {
-        grpcServiceObject.getMetadata((err, metadata, apiResponse_) => {
-          assert.strictEqual(err, error);
-          assert.strictEqual(metadata, null);
-          assert.strictEqual(apiResponse_, apiResponse);
-          done();
-        });
+        assert.strictEqual(grpcServiceObject.getMetadata !== undefined, true);
+        if (grpcServiceObject.getMetadata) {
+          grpcServiceObject.getMetadata(
+              (err: Error, metadata: {}, apiResponse_: {}) => {
+                assert.strictEqual(err, error);
+                assert.strictEqual(metadata, null);
+                assert.strictEqual(apiResponse_, apiResponse);
+                done();
+              });
+        }
       });
     });
 
@@ -143,26 +184,34 @@ describe('GrpcServiceObject', () => {
       const apiResponse = {};
 
       beforeEach(() => {
-        grpcServiceObject.request = (protoOpts, reqOpts, callback) => {
-          callback(null, apiResponse);
-        };
+        grpcServiceObject.request =
+            (protoOpts: {}, reqOpts: {}, callback: Function) => {
+              callback(null, apiResponse);
+            };
       });
 
       it('should exec callback with metadata & API response', done => {
-        grpcServiceObject.getMetadata((err, metadata, apiResponse_) => {
-          assert.ifError(err);
-          assert.strictEqual(metadata, apiResponse);
-          assert.strictEqual(apiResponse_, apiResponse);
-          done();
-        });
+        assert.strictEqual(grpcServiceObject.getMetadata !== undefined, true);
+        if (grpcServiceObject.getMetadata) {
+          grpcServiceObject.getMetadata(
+              (err: Error, metadata: {}, apiResponse_: {}) => {
+                assert.ifError(err);
+                assert.strictEqual(metadata, apiResponse);
+                assert.strictEqual(apiResponse_, apiResponse);
+                done();
+              });
+        }
       });
 
       it('should update the metadata on the instance', done => {
-        grpcServiceObject.getMetadata(err => {
-          assert.ifError(err);
-          assert.strictEqual(grpcServiceObject.metadata, apiResponse);
-          done();
-        });
+        assert.strictEqual(grpcServiceObject.getMetadata !== undefined, true);
+        if (grpcServiceObject.getMetadata) {
+          grpcServiceObject.getMetadata((err: Error) => {
+            assert.ifError(err);
+            assert.strictEqual(grpcServiceObject.metadata, apiResponse);
+            done();
+          });
+        }
       });
     });
   });
@@ -172,27 +221,41 @@ describe('GrpcServiceObject', () => {
     const METADATA = {a: 'c'};
 
     it('should make the correct request', done => {
-      const setMetadataMethod = grpcServiceObject.methods.setMetadata;
+      const setMetadataMethod = grpcServiceObject.methods ?
+          grpcServiceObject.methods.setMetadata :
+          undefined;
       const expectedReqOpts = extend(true, {}, DEFAULT_REQ_OPTS, METADATA);
 
-      grpcServiceObject.methods.setMetadata.reqOpts = DEFAULT_REQ_OPTS;
+      assert.strictEqual(grpcServiceObject.methods !== undefined, true);
+      if (grpcServiceObject.methods) {
+        grpcServiceObject.methods.setMetadata.reqOpts = DEFAULT_REQ_OPTS;
+      }
 
-      grpcServiceObject.request = (protoOpts, reqOpts, callback) => {
-        assert.strictEqual(protoOpts, setMetadataMethod.protoOpts);
-        assert.deepStrictEqual(reqOpts, expectedReqOpts);
-        callback();  // done()
-      };
-
-      grpcServiceObject.setMetadata(METADATA, done);
+      grpcServiceObject.request =
+          (protoOpts: {}, reqOpts: {}, callback: Function) => {
+            assert.strictEqual(setMetadataMethod !== undefined, true);
+            if (setMetadataMethod) {
+              assert.strictEqual(protoOpts, setMetadataMethod.protoOpts);
+            }
+            assert.deepStrictEqual(reqOpts, expectedReqOpts);
+            callback();  // done()
+          };
+      assert.strictEqual(grpcServiceObject.setMetadata !== undefined, true);
+      if (grpcServiceObject.setMetadata) {
+        grpcServiceObject.setMetadata(METADATA, done);
+      }
     });
 
     it('should not require a callback', done => {
-      grpcServiceObject.request = (protoOpts, reqOpts, callback) => {
-        assert.doesNotThrow(callback);
-        done();
-      };
-
-      grpcServiceObject.setMetadata(METADATA);
+      grpcServiceObject.request =
+          (protoOpts: {}, reqOpts: {}, callback: Function) => {
+            assert.doesNotThrow(callback);
+            done();
+          };
+      assert.strictEqual(grpcServiceObject.setMetadata !== undefined, true);
+      if (grpcServiceObject.setMetadata) {
+        grpcServiceObject.setMetadata(METADATA);
+      }
     });
   });
 
@@ -209,8 +272,11 @@ describe('GrpcServiceObject', () => {
         },
       };
 
-      const ret = grpcServiceObject.request.apply(grpcServiceObject, args);
-      assert.strictEqual(ret, expectedReturnValue);
+      assert.strictEqual(grpcServiceObject.request !== undefined, true);
+      if (grpcServiceObject.request) {
+        const ret = grpcServiceObject.request.apply(grpcServiceObject, args);
+        assert.strictEqual(ret, expectedReturnValue);
+      }
     });
   });
 
@@ -227,9 +293,12 @@ describe('GrpcServiceObject', () => {
         },
       };
 
-      const ret =
-          grpcServiceObject.requestStream.apply(grpcServiceObject, args);
-      assert.strictEqual(ret, expectedReturnValue);
+      assert.strictEqual(grpcServiceObject.requestStream !== undefined, true);
+      if (grpcServiceObject.requestStream) {
+        const ret =
+            grpcServiceObject.requestStream.apply(grpcServiceObject, args);
+        assert.strictEqual(ret, expectedReturnValue);
+      }
     });
   });
 
@@ -246,9 +315,13 @@ describe('GrpcServiceObject', () => {
         },
       };
 
-      const ret = grpcServiceObject.requestWritableStream.apply(
-          grpcServiceObject, args);
-      assert.strictEqual(ret, expectedReturnValue);
+      assert.strictEqual(
+          grpcServiceObject.requestWritableStream !== undefined, true);
+      if (grpcServiceObject.requestWritableStream) {
+        const ret = grpcServiceObject.requestWritableStream.apply(
+            grpcServiceObject, args);
+        assert.strictEqual(ret, expectedReturnValue);
+      }
     });
   });
 });
